refactor(config-ui): move plugins/common/sync-policy to pages/blueprint/components
