+++ config-ui/src/pages/blueprint/components/index.ts
@@ -17,3 +17,4 @@
  */
 
 export * from './advanced-editor';
+export * from './sync-policy';
