+++ config-ui/src/pages/blueprint/components/sync-policy/StartFromSelector.tsx
