+++ config-ui/src/pages/blueprint/components/sync-policy/styled.ts
